Disallow difference() between two non-ISO calendars
In the difference() methods of Date and DateTime, check if the calendars
of the two operands are different. If they are different and either one
is ISO, then use the other one. If they are different and both non-ISO,
then throw.

This is because the result of a difference between two dates in
different calendars would be ambiguous as to which calendar the
resulting duration would be expressed in. But as an exception, we allow
one ISO and one non-ISO calendar, for the use case of using an ISO date
as a reference point for calculations.

Closes: #587